Rename router and videos service .js files; move and rename videos data .json file
