+++ src/main.js
@@ -1,5 +1,5 @@
 import { createApp } from 'vue'
-import router from './router/router.js'
+import router from './router'
 import './style.css'
 import App from './App.vue'
